+++ commands/previewPush.go
@@ -11,11 +11,11 @@ import (
 
 	"github.com/urfave/cli/v2"
 
+	"github.com/StackExchange/dnscontrol/v3/internal/printer"
 	"github.com/StackExchange/dnscontrol/v3/models"
 	"github.com/StackExchange/dnscontrol/v3/pkg/credsfile"
 	"github.com/StackExchange/dnscontrol/v3/pkg/normalize"
 	"github.com/StackExchange/dnscontrol/v3/pkg/notifications"
-	"github.com/StackExchange/dnscontrol/v3/pkg/printer"
 	"github.com/StackExchange/dnscontrol/v3/providers"
 )
