+++ internal/printer/console.go
