+++ pkg/diff/diff.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/gobwas/glob"
 
+	"github.com/StackExchange/dnscontrol/v3/internal/printer"
 	"github.com/StackExchange/dnscontrol/v3/models"
-	"github.com/StackExchange/dnscontrol/v3/pkg/printer"
 )
 
 // Correlation stores a difference between two domains.
